+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 # Changelog
 
-## 0.3.9 - 2025-07-14
+## 0.4 - 2025-07-14 14:00 UTC
+- added drag and drop reordering for books and sub-folders
+- simplified documentation and added help link
+- improved Docker instructions
+
+## 0.3.9 - 2025-07-14 13:58 UTC
 - added changelog page and link
 - bold book names and remember sidebar collapse state
 - books can be reordered from the home page
 - moved author display and cleaned sidebar headings
 - added list buttons in chapter editor
 
-## 0.3.8 - 2025-07-14
+## 0.3.8 - 2025-07-14 13:00 UTC
 - fix folder view for sub-folders
 - removed broken copy buttons